requirements.txt
Flask==3.0.2
Werkzeug==3.0.2
Jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.1.0
pdfminer.six==20231228
pandas==2.2.2
transformers==4.40.2
torch==2.3.0
